Moved tests for JSON signature to separate module
